% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_weighted_auc.R
\name{wAUC}
\alias{wAUC}
\alias{wAUC_resampling}
\alias{wAUC_exact}
\title{Weighted c-statistic}
\usage{
wAUC(y, p, w, na.rm = TRUE, rescale.w = FALSE, method = "resampling", ...)
wAUC_resampling(y, p, w, na.rm = TRUE, rescale.w = FALSE, replace = TRUE,
                size = length(p), I = 1000, level = .95, nonunity = FALSE,
                nonzero = FALSE, ret.resamples = FALSE,
                AUC_method = "default", ...)
wAUC_exact(y, p, w, na.rm = TRUE, rescale.w = FALSE, nonunity = FALSE,
           nonzero = FALSE,  ...)
}
\arguments{
\item{y}{numeric outcome vector: each value 1 or 0}

\item{p}{numeric predicted probabilities vector: 0 <= p <= 1}

\item{w}{numeric weights. w >= for exact, 0 <= w <= 1 for resampling}

\item{na.rm}{Should NA's be removed?}

\item{rescale.w}{Should w be rescaled to 0 <= w <= 1?}

\item{method}{Use the "resampling" or "exact" method for calculation?}

\item{...}{Passed to \code{wAUC_resampling} and \code{wAUC_exact}.}

\item{replace}{Should sample be performed with replacement?}

\item{size}{If sampling is performed with replacement, how many observations should be sampled per
propensity weighted bootstrap?}

\item{I}{number of resamples}

\item{level}{Confidence level or percentile for the propensity weighted bootstrap percentiles with
replacement.}

\item{ret.resamples}{Should all the resamples be returned?}

\item{AUC_method}{default or factorial. default is always faster.}

\item{nonunity}{Should a correction be applied when the estimate is exactly one?}

\item{nonzero}{Should a correction be applied when the estimate is exactly zero?}
}
\value{
A list of class wAUC. The exact method contains only a scalar: the AUC / c-statistic. The
resampling version contains a list with the statistics and possibly the resamples.
}
\description{
Calculate a weighted AUC or c-statistic, exactly or by resampling
}
\details{
Perfect (c = 1) and perfectly wrong (c = 0) values may lead to computational issues when
(meta-)analyzed. Set nonunity to nonzero to TRUE to avoid this, which then respectively substracts or
adds a value equal to one tie when this is the case.

For the resampling methods, setting replace = TRUE essentially gives a weighted bootstrap
estimate. Hence, a confidence interval is returned. Setting replace = FALSE, yields smaller
samples for any w < 1, but returns the exact same same for all w == 1. In the latter case
the percentiles have the same value as the point estimate and cannot be interpreted as
confidence intervals. Hence, no ci is returned when replace = FALSE.

Note that unique_events and unique_nonevents may be switched. No check is performed for this.
}
\author{
Valentijn de Jong <Valentijn.M.T.de.Jong@gmail.com>
}
